requirements.txt
adaptix==3.0.0b11
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
click==8.2.1
dishka==1.6.0
dynaconf==3.2.11
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
iniconfig==2.1.0
mako==1.3.10
markupsafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-core==2.33.2
pyjwt==2.10.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.1
uvicorn==0.34.2
